# web/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SchedulePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScheduleView"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "SchedulePage",
        "type": "object"
      },
      "SchedulePatch": {
        "additionalProperties": false,
        "description": "Partial schedule updates supported by PATCH.",
        "properties": {
          "cron": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cron"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "priority": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          }
        },
        "title": "SchedulePatch",
        "type": "object"
      },
...
      "ScheduleRunRequest": {
        "additionalProperties": false,
        "properties": {
          "priority": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          }
        },
        "title": "ScheduleRunRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "MaaAssistantArknights 的 Web 服务（重构版）。\n\n- **鉴权**：单一 `access_token`，四渠道按优先级取值 —— `Authorization: Bearer` → `X-Token` → `?token=` → cookie `maa_token`；未配置 `access_token` 时全部端点免鉴权\n- **错误体**：所有非 2xx 响应都是 `{\"error\": {\"code\", \"message\", \"details\"}}`，`code` 是可枚举的错误码\n- **尾斜杠**：全部路由不带尾斜杠，带尾斜杠一律 404（`redirect_slashes=False`）\n- `ws` 与 `mcp` 分组只在本文档占位说明协议与鉴权方式，不能在 Swagger UI 中调试",
    "title": "MAA-API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/pipelines": {
      "get": {
        "operationId": "pipelines_list_pipelines",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "schedule_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Schedule Id"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Pipelines List Pipelines",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_PARAMETER",
                    "details": {},
                    "message": "请求不成立"
                  }
                }
              }
            },
            "description": "请求不成立：INVALID_PARAMETER、INVALID_PAGINATION"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "details": {},
                    "message": "token 缺失或不匹配"
                  }
                }
              }
            },
            "description": "token 缺失或不匹配：UNAUTHORIZED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询流水线历史",
        "tags": [
          "pipelines"
        ]
      },
      "post": {
        "operationId": "pipelines_submit_pipeline",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Pipelines Submit Pipeline",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "PIPELINE_EMPTY",
                    "details": {},
                    "message": "请求不成立"
                  }
                }
              }
            },
            "description": "请求不成立：PIPELINE_EMPTY、PIPELINE_TOO_MANY_TASKS、UNKNOWN_TASK_TYPE"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "details": {},
                    "message": "token 缺失或不匹配"
                  }
                }
              }
            },
            "description": "token 缺失或不匹配：UNAUTHORIZED"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "QUEUE_PAUSED",
                    "details": {},
                    "message": "与当前状态冲突"
                  }
                }
              }
            },
            "description": "与当前状态冲突：QUEUE_PAUSED、IDEMPOTENCY_KEY_CONFLICT"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "TASK_PARAM_INVALID",
                    "details": {},
                    "message": "参数校验失败"
                  }
                }
              }
            },
            "description": "参数校验失败：TASK_PARAM_INVALID"
          },
          "429": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "QUEUE_FULL",
                    "details": {},
                    "message": "请求过于频繁"
                  }
                }
              }
            },
            "description": "请求过于频繁：QUEUE_FULL",
            "headers": {
              "Retry-After": {
                "description": "建议的重试等待秒数（整数）",
                "schema": {
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "提交流水线",
        "tags": [
          "pipelines"
        ]
      }
    },
...
    "/api/pipelines/{pipeline_id}": {
      "delete": {
        "operationId": "pipelines_cancel_pipeline",
        "parameters": [
          {
            "in": "path",
            "name": "pipeline_id",
            "required": true,
            "schema": {
              "title": "Pipeline Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Pipelines Cancel Pipeline",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "details": {},
                    "message": "token 缺失或不匹配"
                  }
                }
              }
            },
            "description": "token 缺失或不匹配：UNAUTHORIZED"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "PIPELINE_NOT_FOUND",
                    "details": {},
                    "message": "资源不存在"
                  }
                }
              }
            },
            "description": "资源不存在：PIPELINE_NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "PIPELINE_NOT_CANCELLABLE",
                    "details": {},
                    "message": "与当前状态冲突"
                  }
                }
              }
            },
            "description": "与当前状态冲突：PIPELINE_NOT_CANCELLABLE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "取消流水线",
        "tags": [
          "pipelines"
        ]
      },
      "get": {
        "operationId": "pipelines_get_pipeline",
        "parameters": [
          {
            "in": "path",
            "name": "pipeline_id",
            "required": true,
            "schema": {
              "title": "Pipeline Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Pipelines Get Pipeline",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "details": {},
                    "message": "token 缺失或不匹配"
                  }
                }
              }
            },
            "description": "token 缺失或不匹配：UNAUTHORIZED"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "PIPELINE_NOT_FOUND",
                    "details": {},
                    "message": "资源不存在"
                  }
                }
              }
            },
            "description": "资源不存在：PIPELINE_NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "获取流水线详情",
        "tags": [
          "pipelines"
        ]
      }
    },
...
    "/api/schedules": {
      "get": {
        "operationId": "schedules_list_schedules",
        "parameters": [
          {
            "in": "query",
            "name": "enabled",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Enabled"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "details": {},
                    "message": "token 缺失或不匹配"
                  }
                }
              }
            },
            "description": "token 缺失或不匹配：UNAUTHORIZED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "读取定时任务",
        "tags": [
          "schedules"
        ]
      },
      "post": {
        "operationId": "schedules_create_schedule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SCHEDULE_CRON_INVALID",
                    "details": {},
                    "message": "请求不成立"
                  }
                }
              }
            },
            "description": "请求不成立：SCHEDULE_CRON_INVALID、UNKNOWN_TASK_TYPE、TASK_PARAM_DEPRECATED"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "details": {},
                    "message": "token 缺失或不匹配"
                  }
                }
              }
            },
            "description": "token 缺失或不匹配：UNAUTHORIZED"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SCHEDULE_NAME_CONFLICT",
                    "details": {},
                    "message": "与当前状态冲突"
                  }
                }
              }
            },
            "description": "与当前状态冲突：SCHEDULE_NAME_CONFLICT"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "TASK_PARAM_INVALID",
                    "details": {},
                    "message": "参数校验失败"
                  }
                }
              }
            },
            "description": "参数校验失败：TASK_PARAM_INVALID、VALIDATION_ERROR"
          }
        },
        "summary": "新建定时任务",
        "tags": [
          "schedules"
        ]
      }
    },
    "/api/schedules/{schedule_id}": {
      "delete": {
        "operationId": "schedules_delete_schedule",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "details": {},
                    "message": "token 缺失或不匹配"
                  }
                }
              }
            },
            "description": "token 缺失或不匹配：UNAUTHORIZED"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SCHEDULE_NOT_FOUND",
                    "details": {},
                    "message": "资源不存在"
                  }
                }
              }
            },
            "description": "资源不存在：SCHEDULE_NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "删除定时任务",
        "tags": [
          "schedules"
        ]
      },
      "get": {
        "operationId": "schedules_get_schedule",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "details": {},
                    "message": "token 缺失或不匹配"
                  }
                }
              }
            },
            "description": "token 缺失或不匹配：UNAUTHORIZED"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SCHEDULE_NOT_FOUND",
                    "details": {},
                    "message": "资源不存在"
                  }
                }
              }
            },
            "description": "资源不存在：SCHEDULE_NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "读取定时任务与近期执行结果",
        "tags": [
          "schedules"
        ]
      },
      "patch": {
        "operationId": "schedules_patch_schedule",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchedulePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SCHEDULE_CRON_INVALID",
                    "details": {},
                    "message": "请求不成立"
                  }
                }
              }
            },
            "description": "请求不成立：SCHEDULE_CRON_INVALID"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "details": {},
                    "message": "token 缺失或不匹配"
                  }
                }
              }
            },
            "description": "token 缺失或不匹配：UNAUTHORIZED"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SCHEDULE_NOT_FOUND",
                    "details": {},
                    "message": "资源不存在"
                  }
                }
              }
            },
            "description": "资源不存在：SCHEDULE_NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "details": {},
                    "message": "参数校验失败"
                  }
                }
              }
            },
            "description": "参数校验失败：VALIDATION_ERROR"
          }
        },
        "summary": "局部更新定时任务",
        "tags": [
          "schedules"
        ]
      },
      "put": {
        "operationId": "schedules_update_schedule",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SCHEDULE_CRON_INVALID",
                    "details": {},
                    "message": "请求不成立"
                  }
                }
              }
            },
            "description": "请求不成立：SCHEDULE_CRON_INVALID、UNKNOWN_TASK_TYPE、TASK_PARAM_DEPRECATED"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "details": {},
                    "message": "token 缺失或不匹配"
                  }
                }
              }
            },
            "description": "token 缺失或不匹配：UNAUTHORIZED"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SCHEDULE_NOT_FOUND",
                    "details": {},
                    "message": "资源不存在"
                  }
                }
              }
            },
            "description": "资源不存在：SCHEDULE_NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SCHEDULE_NAME_CONFLICT",
                    "details": {},
                    "message": "与当前状态冲突"
                  }
                }
              }
            },
            "description": "与当前状态冲突：SCHEDULE_NAME_CONFLICT"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "TASK_PARAM_INVALID",
                    "details": {},
                    "message": "参数校验失败"
                  }
                }
              }
            },
            "description": "参数校验失败：TASK_PARAM_INVALID、VALIDATION_ERROR"
          }
        },
        "summary": "全量替换定时任务",
        "tags": [
          "schedules"
        ]
      }
    },
    "/api/schedules/{schedule_id}/run": {
      "post": {
        "operationId": "schedules_run_schedule",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleRunAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "PIPELINE_EMPTY",
                    "details": {},
                    "message": "请求不成立"
                  }
                }
              }
            },
            "description": "请求不成立：PIPELINE_EMPTY、UNKNOWN_TASK_TYPE"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "details": {},
                    "message": "token 缺失或不匹配"
                  }
                }
              }
            },
            "description": "token 缺失或不匹配：UNAUTHORIZED"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SCHEDULE_NOT_FOUND",
                    "details": {},
                    "message": "资源不存在"
                  }
                }
              }
            },
            "description": "资源不存在：SCHEDULE_NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "QUEUE_PAUSED",
                    "details": {},
                    "message": "与当前状态冲突"
                  }
                }
              }
            },
            "description": "与当前状态冲突：QUEUE_PAUSED"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "TASK_PARAM_INVALID",
                    "details": {},
                    "message": "参数校验失败"
                  }
                }
              }
            },
            "description": "参数校验失败：TASK_PARAM_INVALID"
          },
          "429": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "QUEUE_FULL",
                    "details": {},
                    "message": "请求过于频繁"
                  }
                }
              }
            },
            "description": "请求过于频繁：QUEUE_FULL",
            "headers": {
              "Retry-After": {
                "description": "建议的重试等待秒数（整数）",
                "schema": {
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "立即运行定时任务",
        "tags": [
          "schedules"
        ]
      }
    },
...
  },
...
}